fixed issue where feed was sorted by time at day granularity not seconds granularity

+++ recommender/FeedRecommender.ts
@@ -66,7 +66,7 @@ export default class FeedRecommender {
 			WHERE 
 				strategy_name = :stratName
 			ORDER BY
-				(EXTRACT(EPOCH FROM (CURRENT_TIMESTAMP - created_at)) / (60 * 60 * 24))::integer ASC,
+				(EXTRACT(EPOCH FROM (CURRENT_TIMESTAMP - created_at)) / (60 * 60 * 24))::numeric ASC,
 				v DESC
 			LIMIT :limit OFFSET :offset;
 		`, { stratName, limit, offset })
